+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [0.2] - 2019-4-29
+## [0.5] - 2019-10-28
 ### Added
 - `version_check.py`
 - `CODE_OF_CONDUCT.md`
@@ -13,7 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `PULL_REQUEST_TEMPLATE.md`
 - `build_exe.bat`
 - `build_unix.sh`
+- `GOPEM.iss`
+- Check update button
+- Save plot button
 - Report generating controller
+- Console print controller
 - PyPI package
 ### Changed
 - `CONTRIBUTING.md` modified
@@ -22,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Menu bug in `Windows` fixed
 - Padulles-Amphlett model default test case bug fixed
 - Docstrings modified
+- Report default location bug fixed
+- Icon bug fixed
 ## [0.1] - 2018-9-10
 ### Added
 - GUI and plotter
@@ -29,6 +35,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Documents and README
 
 
-[Unreleased]: https://github.com/ECSIM/gopem/compare/v0.2...develop
-[0.2]: https://github.com/ECSIM/gopem/compare/v0.1...v0.2
+[Unreleased]: https://github.com/ECSIM/gopem/compare/v0.5...develop
+[0.5]: https://github.com/ECSIM/gopem/compare/v0.1...v0.5
 [0.1]: https://github.com/ECSIM/gopem/compare/0f35fc9...v0.1